refactor: extract TransactionsView and CopyButton into their own components
TaxSummary was getting long. Pull the per-transaction table into
TransactionsView.tsx, and the small Copy button (shared with the totals-view
CraBlock) into CopyButton.tsx.
